FROMLIST: qcom-tgu: Add TGU decode support
Decoding is when all the potential pieces for creating a trigger
are brought together for a given step. Example - there may be a
counter keeping track of some occurrences and a priority-group that
is being used to detect a pattern on the sense inputs. These 2
inputs to condition_decode must be programmed, for a given step,
to establish the condition for the trigger, or movement to another
steps.

Link: https://lore.kernel.org/all/20251203090055.2432719-5-songwei.chai@oss.qualcomm.com/
Signed-off-by: Songwei Chai <songwei.chai@oss.qualcomm.com>